refactor: modularize dashboard
